pyproject.toml
[project]
name = "nk-autocode"
dynamic = ["version"]
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.0,<4.0.0"
dependencies = [
    "openai>=1.82.0",
    "pillow>=11.3.0",
    "python-dotenv>=1.1.0",
]
license = { text = "MIT" }

[dependency-groups]
dev = [
    "ipython>=9.0.2",
    "mypy>=1.15.0",
    "pdbpp>=0.10.3",
    "pytest>=8.3.5",
    "ruff>=0.11.4",
]

[build-system]
requires = ["hatchling", "uv-dynamic-versioning"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["nk_autocode", "nk_autodoc"]

[tool.hatch.version]
source = "uv-dynamic-versioning"

[tool.uv-dynamic-versioning]
vcs = "git"
style = "pep440"
bump = true
fallback-version = "0.0.0"

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["T201", "COM812", "ISC001"]
fixable = ["ALL"]
unfixable = ["F841"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = true

[tool.ruff.lint.pylint]
max-args = 6

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["D", "S101", "N802", "ARG"]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.12"
warn_unused_configs = true
disallow_untyped_defs = true
show_error_codes = true
no_implicit_optional = true
warn_return_any = true
warn_unused_ignores = true
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["tests.*"]
disallow_untyped_defs = false

[[tool.mypy.overrides]]
module = ["samples.*"]
ignore_errors = true
